Implement fancy command
Fancy command is a widget pipe which pipes output from a command
thru this widget into another widget.
This allows to transform the ouptut in this widget like syntax highlight
or anything else.

The first implementation is using chroma  just for the testing.

Also attach command was updated to deal with `fancy` keyword to
pipe correctly the output.