Implement suspend/resume gocui
Vim and other full terminal application needs full access to inputs
from terminal.
When gocui is running, it can steal some of the inputs for Vim.
To overcome this, we need to close gocui.

To be able to suspend/close gocui and resume it later, two channels
were added and new error which signal supend request.

When gocui.MainLoop() ends with suspend error, we would close gocui and
signal thru one channel the closure.
Afterwards we would wait for signal to resume on other channel.
When resume signaled, we would reinitialize gocui with all the widgets
it had setup.

Widget has `body` variable which stores the content of `Print` in it.
This way we can always display original content after resume.
Views are reinitialized so they are empty after gocui init.